vinci-inference/requirements/client.txt
opencv-python-headless==4.5.5.64
decord==0.6.0
numpy==1.26.4
requests
pillow
# sseclient
sseclient-py